fix: refactor netjsongraph.core.js && test passed
