+++ src/js/netjsongraph.render.js
@@ -5,6 +5,7 @@ import "echarts/lib/chart/graph";
 import "echarts/lib/component/tooltip";
 import "echarts/lib/component/title";
 import "echarts/lib/component/toolbox";
+import "echarts/lib/component/legend";
 import "zrender/lib/svg/svg";
 
 import "../../lib/js/leaflet.min.js";
